Add support for RESTful URLs
Nereid URLs used a selection box to choose the methods accepted
by the handler. This limited GET, POST and both GET & POST as
the available options. This patch replaces the selection field
with checkboxes for different HTTP methods. The patch gracefuly
converst existing URLs written with methods selection into the
new format. The old XML is still valid, but will be deprecated
in the upcoming release