Remove dependency on ccy #41
CCY was used to identify the currency from the
country part of the locale (eg. US in  en_US) to
identify the currency. The library is both a burden
and a pain to install automatically.

This patch removes the dependency on the library
and adds a new field to language where the default
currency for the language can be specified. If a default
currency is not specified in the language, it is picked up
from the currency of the company of the website